Split off a library
