+++ src/main.rs
@@ -1,13 +1,11 @@
 #![no_std]
 #![no_main]
 #![feature(custom_test_frameworks)]
-#![test_runner(crate::test_runner)]
+#![test_runner(blog_os::test_runner)]
 #![reexport_test_harness_main = "test_main"]
 
+use blog_os::println;
 use core::panic::PanicInfo;
-
-mod serial;
-mod vga_buffer;
 
 #[no_mangle]
 pub extern "C" fn _start() -> ! {
@@ -17,15 +15,6 @@ pub extern "C" fn _start() -> ! {
     test_main();
 
     loop {}
-}
-
-#[cfg(test)]
-fn test_runner(tests: &[&dyn Fn()]) {
-    serial_println!("Running {} tests", tests.len());
-    for test in tests {
-        test();
-    }
-    unsafe { exit_qemu(QemuExitCode::Success) };
 }
 
 /// This function is called on panic.
@@ -39,24 +28,5 @@ fn panic(info: &PanicInfo) -> ! {
 #[cfg(test)]
 #[panic_handler]
 fn panic(info: &PanicInfo) -> ! {
-    serial_println!("[failed]\n");
-    serial_println!("Error: {}\n", info);
-    unsafe {
-        exit_qemu(QemuExitCode::Failed);
-    }
-    loop {}
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[repr(u32)]
-pub enum QemuExitCode {
-    Success = 0x10,
-    Failed = 0x11,
-}
-
-pub unsafe fn exit_qemu(exit_code: QemuExitCode) {
-    use x86_64::instructions::port::Port;
-
-    let mut port = Port::new(0xf4);
-    port.write(exit_code as u32);
+    blog_os::test_panic_handler(info)
 }
